+++ packages/ag-charts-enterprise/src/box-plot/boxPlotSeries.ts
@@ -1,5 +1,7 @@
-import type { AgCartesianSeriesTooltipRendererParams } from 'ag-charts-community';
-import { _ModuleSupport, _Scene } from 'ag-charts-community';
+import type { AgCartesianSeriesTooltipRendererParams, _Scene } from 'ag-charts-community';
+import { _ModuleSupport } from 'ag-charts-community';
+import type { BoxPlotNodeDatum } from './boxPlotTypes';
+import { BoxPlotGroup } from './boxPlotGroup';
 
 const {
     CartesianSeries,
@@ -15,36 +17,10 @@ const {
     valueProperty,
 } = _ModuleSupport;
 
-enum GroupTags {
-    Box,
-    Median,
-    Outline,
-    Whisker,
-    Cap,
-}
-
-interface BoxPlotNodeDatum extends _ModuleSupport.CartesianSeriesNodeDatum {
-    readonly index: number;
-    readonly xValue: number;
-    readonly bandwidth: number;
-
-    readonly yValue: number;
-    readonly minValue: number;
-    readonly q1Value: number;
-    readonly medianValue: number;
-    readonly q3Value: number;
-    readonly maxValue: number;
-
-    readonly fill: string;
-    readonly fillOpacity: number;
-    readonly stroke: string;
-    readonly strokeWidth: number;
-    readonly strokeOpacity: number;
-    readonly lineDash: number[];
-    readonly lineDashOffset: number;
-}
-
-export class BoxPlotSeries extends CartesianSeries<_ModuleSupport.SeriesNodeDataContext<BoxPlotNodeDatum>> {
+export class BoxPlotSeries extends CartesianSeries<
+    _ModuleSupport.SeriesNodeDataContext<BoxPlotNodeDatum>,
+    BoxPlotGroup
+> {
     @Validate(OPT_STRING)
     xKey?: string = undefined;
 
@@ -250,7 +226,7 @@ export class BoxPlotSeries extends CartesianSeries<_ModuleSupport.SeriesNodeData
 
     protected async updateDatumSelection(opts: {
         nodeData: BoxPlotNodeDatum[];
-        datumSelection: _Scene.Selection<_Scene.Group, BoxPlotNodeDatum>;
+        datumSelection: _Scene.Selection<BoxPlotGroup, BoxPlotNodeDatum>;
     }) {
         const { nodeData, datumSelection } = opts;
         const data = nodeData ?? [];
@@ -258,92 +234,11 @@ export class BoxPlotSeries extends CartesianSeries<_ModuleSupport.SeriesNodeData
     }
 
     protected async updateDatumNodes(opts: {
-        datumSelection: _Scene.Selection<_Scene.Group, BoxPlotNodeDatum>;
+        datumSelection: _Scene.Selection<BoxPlotGroup, BoxPlotNodeDatum>;
         isHighlight: boolean;
     }) {
-        opts.datumSelection.each((group, datum) => {
-            const {
-                bandwidth,
-                minValue,
-                q1Value,
-                medianValue,
-                q3Value,
-                maxValue,
-                yValue,
-                fill,
-                fillOpacity,
-                stroke,
-                strokeWidth,
-                strokeOpacity,
-                lineDash,
-                lineDashOffset,
-            } = datum;
-
-            const selection = _Scene.Selection.select(group, _Scene.Rect);
-            const boxes = selection.selectByTag<_Scene.Rect>(GroupTags.Box);
-            const [outline] = selection.selectByTag<_Scene.Rect>(GroupTags.Outline);
-            const [median] = selection.selectByTag<_Scene.Line>(GroupTags.Median);
-            const whiskers = selection.selectByTag<_Scene.Line>(GroupTags.Whisker);
-            const caps = selection.selectByTag<_Scene.Line>(GroupTags.Cap);
-
-            outline.setStyles({ x: q1Value, y: yValue, width: q3Value - q1Value, height: bandwidth });
-
-            boxes[0].setStyles({
-                x: q1Value,
-                y: yValue,
-                width: Math.round(medianValue - q1Value + strokeWidth / 2),
-                height: bandwidth,
-            });
-
-            boxes[1].setStyles({
-                x: Math.round(medianValue - strokeWidth / 2),
-                y: yValue,
-                width: Math.floor(q3Value - medianValue + strokeWidth / 2),
-                height: bandwidth,
-            });
-
-            median.setStyles({
-                x: medianValue,
-                y1: yValue + strokeWidth,
-                y2: yValue + bandwidth - strokeWidth,
-            });
-
-            const capSize = 0.5;
-            const capY1 = yValue + bandwidth * capSize * 0.5;
-            const capY2 = yValue + bandwidth * capSize * 1.5;
-
-            caps[0].setStyles({ x: minValue, y1: capY1, y2: capY2 });
-            caps[1].setStyles({ x: maxValue, y1: capY1, y2: capY2 });
-
-            whiskers[0].setStyles({
-                x1: Math.round(minValue + strokeWidth / 2),
-                x2: q1Value,
-                y: yValue + bandwidth * 0.5,
-            });
-
-            whiskers[1].setStyles({
-                x1: q3Value,
-                x2: Math.round(maxValue - strokeWidth / 2),
-                y: yValue + bandwidth * 0.5,
-            });
-
-            // fill only elements
-            for (const element of boxes) {
-                element.fill = fill;
-                element.fillOpacity = fillOpacity;
-                element.strokeWidth = strokeWidth * 2;
-                element.strokeOpacity = 0;
-            }
-
-            // stroke only elements
-            for (const element of [outline, median, ...whiskers, ...caps]) {
-                element.fillOpacity = 0;
-                element.stroke = stroke;
-                element.strokeWidth = strokeWidth;
-                element.strokeOpacity = strokeOpacity;
-                element.lineDash = lineDash;
-                element.lineDashOffset = lineDashOffset;
-            }
+        opts.datumSelection.each((boxPlot, datum) => {
+            boxPlot.updateDatumStyles(datum);
         });
     }
 
@@ -363,18 +258,7 @@ export class BoxPlotSeries extends CartesianSeries<_ModuleSupport.SeriesNodeData
     }
 
     protected nodeFactory() {
-        const group = new _Scene.Group({ layer: true });
-        group.append([
-            new _Scene.Rect({ tag: GroupTags.Box }),
-            new _Scene.Rect({ tag: GroupTags.Box }),
-            new _Scene.Rect({ tag: GroupTags.Outline }),
-            new _Scene.Line({ tag: GroupTags.Median }),
-            new _Scene.Line({ tag: GroupTags.Whisker }),
-            new _Scene.Line({ tag: GroupTags.Whisker }),
-            new _Scene.Line({ tag: GroupTags.Cap }),
-            new _Scene.Line({ tag: GroupTags.Cap }),
-        ]);
-        return group;
+        return new BoxPlotGroup();
     }
 
     getBandScalePadding() {
