+++ packages/ag-charts-enterprise/src/box-plot/boxPlotTypes.ts
@@ -1,0 +1,21 @@
+import type { _ModuleSupport } from 'ag-charts-community';
+export interface BoxPlotNodeDatum extends _ModuleSupport.CartesianSeriesNodeDatum {
+    readonly index: number;
+    readonly xValue: number;
+    readonly bandwidth: number;
+
+    readonly yValue: number;
+    readonly minValue: number;
+    readonly q1Value: number;
+    readonly medianValue: number;
+    readonly q3Value: number;
+    readonly maxValue: number;
+
+    readonly fill: string;
+    readonly fillOpacity: number;
+    readonly stroke: string;
+    readonly strokeWidth: number;
+    readonly strokeOpacity: number;
+    readonly lineDash: number[];
+    readonly lineDashOffset: number;
+}
